Settings query params dto.

+++ src/server/models/substation.rs
@@ -12,8 +12,8 @@ pub struct PostSubstationSDto{
 #[derive(Debug, Serialize, Deserialize)]
 pub struct PutSubstationSDto{
     pub code: String, // unique code of substation
-    pub latitude: u64,
-    pub longitude: u64,
+    pub latitude: f64,
+    pub longitude: f64,
     pub max_power: i64
 }
 
@@ -21,8 +21,8 @@ pub struct PutSubstationSDto{
 pub struct SubstationSDto{
     pub id: String,
     pub code: String, // unique code of substation
-    pub latitude: u64,
-    pub longitude: u64,
-    pub max_power: i64
+    pub latitude: f64,
+    pub longitude: f64,
+    pub max_power: u64
 }
